fix(build): check if static dir exists

+++ gridsome/lib/build.js
@@ -19,7 +19,10 @@ module.exports = async (context, args) => {
 
   await plugins.callHook('beforeBuild', { context, config })
   await fs.remove(config.outDir)
-  await fs.copy(config.staticDir, config.targetDir)
+
+  if (fs.existsSync(config.staticDir)) {
+    await fs.copy(config.staticDir, config.targetDir)
+  }
 
   const queue = await createRenderQueue(app)
